+++ src/components/auth/intro/IntroLogo.tsx
@@ -1,6 +1,6 @@
 import SymbolWhite from "@/assets/logo/symbol-white.svg?react";
 
-export default function IntroSlide1({ isActive }: { isActive: boolean }) {
+export default function IntroLogo({ isActive }: { isActive: boolean }) {
   return (
     <div
       className={`absolute inset-0 flex flex-col items-center justify-center transition-opacity duration-1000 ease-in-out bg-linear-to-b from-logo-1 to-logo-2 ${
